Added opening hours property editor
Based on the similar property editor from https://github.com/umbraco-community/Our.Umbraco.OpeningHours - most of that package's current code is from when I was lead dev on the package and where I rewrote most of the property editor. And the package is MIT as well, so shouldn't be a problem.

By moving/copying the code to Limbo Time, I can hopefully take more ownership of the code as well a commit more company time to maintaining the package.

The code in this commit also contains unreleased progress compared to the original repo that I made for Skybrud Time - before rebranding the package as Limbo Time.